Implement visibility control for map markers
This update introduces a `show` property to the marker data structure in the `markers.json` file and conditions rendering of markers on its associated value in `ProjectMap.vue`. This change lets us selectively display markers on the volunteering Ukraine map which enhances the code's flexibility for possible feature additions or changes in the future.